+++ packages/oruga/src/composables/unrefElement.ts
@@ -24,14 +24,22 @@ export type UnRefElementReturn<T extends MaybeElement = MaybeElement> =
           ? HTMLElement
           : T | undefined;
 
-/**
- * Get the dom element of a ref of element or Vue component instance
- *
- * @param elRef
- */
+/** Get the dom element of a ref of element or Vue component instance */
 export function unrefElement<T extends MaybeElement>(
     elRef: MaybeRefOrGetter<T> | MaybeRef<T>,
 ): UnRefElementReturn<T> {
     const plain = toValue(elRef);
     return (plain as ComponentPublicInstance)?.$el ?? plain;
+}
+
+/** Resolve an HTML element based on query selector or an explizit dom element */
+export function resolveElement(
+    target: MaybeRefOrGetter<HTMLElement | string>,
+): HTMLElement | null {
+    const targetQuery = toValue(target);
+    // query element if target is a string
+    if (typeof targetQuery === "string")
+        return document.querySelector<HTMLElement>(targetQuery);
+    // else unwrap element
+    else return unrefElement(targetQuery);
 }
